create first running version doing nothing
